fintech_app/requirements.txt
flask
psycopg2-binary
pandas
scikit-learn
apache-airflow==2.7.3
